requirements.txt
Flask==2.0.1
redis==3.5.3
rq==1.10.0
psycopg2-binary==2.9.1
requests==2.26.0